fix: rename PropertyReservationsPaginator and PropertyReservationStream
